Fix kill test & add Semaphore
Replace mutex by semaphore, so that we easily handle switch between parallel
and sequential execution. Furthermore, the maximal amount of concurrent
test runs is limited to the number of cores. So that system is still
usable.

Signal emited to kill a test could be received by any waiting thread.
Now added a wrapper that is created for a specific test, so that signal
always is emited for right one.

Extract tearDown procedure to shared lambda function.

Adapt killTest procedure, to try terminating a few times before hard
kill. Before the TestDriver actually reads for the signal, there is the
RunEnv.bat setput, python setup where the signal would get lost otherwise.
Especially when "Run all Tests" and "Kill all Tests" is used in combination.

Add an early check for kill request after semaphore acquire, so that test
does not start if it already should be killed.