Adapt CMakeBuildConfig handling
Only add CMakeBuildPrefix if there is more than one to the NameRole.
So make NameRole explicit, but don't use if for xmlPath to have the same
name if a new CMakeBuildConfig is added or removed.

Change autoRun interpretation. Only run tests during start-up if there
was a result before. So we don't autoRun on first start as the user has
no way to stop the running there.

Also if there are more than one CMakeBuildConfigs, only autoRun one,
as concurrent running of same test is not supported yet. (They would
overwrite the test data output dir / LatestLog).