Add testsController
to manage multiple tests at once.
Save the "latest" folder only for the last execution and not for all.
This safes memory, as some tests have very large data folders.

Save the output xml in the parent dir and append the time to distinguish them.

There still open TODOs left to get all the functionality as before.

Also reformated some files which where way of.